Refactoring backend routes code
